change the algorithm of assining zone when zone=0.
If origZone (or destZone) in a packet header is equal to zero, then
first look for a link with net coinciding with the origNet (or destNet).
If such link is found, assign its zone to origZone (or destZone).
Otherwise, look among the configuration owner addresses for the net
coinciding with the origNet (or destNet). If such an address is found,
assign its zone to origZone (or destZone). Otherwise, assign the main
address zone to origZone (or destZone).